refactor(intellij): get better performance and refactor notifications
